requirements.txt
facebook_business
google-analytics-data
google-api-python-client
google-auth-oauthlib
google-ads
pandas
numpy
apscheduler
flask